problems with events
